requirements.txt
streamlit
torch
librosa
matplotlib
pyaudio
sounddevice
transformers
openai-whisper
scipy
